+++ app/_components/pdf-viewer.tsx
@@ -15,9 +15,12 @@ import CustomRenderer from "@/_components/pdf-custom-renderer";
 import { getPageNumbers } from "@/_lib/get-page-numbers";
 import { PDF_TO_CSS_UNITS } from "@/_lib/pixels-per-inch";
 import { RenderContext } from "@/_hooks/use-render-context";
+import { useTransformerContext } from "@/_hooks/use-transform-context";
 import { StitchSettings } from "@/_lib/interfaces/stitch-settings";
 import { StitchSettingsAction } from "@/_reducers/stitchSettingsReducer";
-import { Layers, getLayersFromPdf } from "@/_lib/layers";
+import { getLayersFromPdf, Layers } from "@/_lib/layers";
+import { LoadStatusEnum } from "@/_lib/load-status-enum";
+import { Point } from "@/_lib/point";
 
 pdfjs.GlobalWorkerOptions.workerSrc = new URL(
   "pdfjs-dist/build/pdf.worker.min.js",
@@ -37,6 +40,9 @@ export default function PdfViewer({
   stitchSettings,
   filter,
   magnifying,
+  setPdfLoadStatus,
+  setLineThicknessStatus,
+  gridCenter,
 }: {
   file: any;
   layers: Layers;
@@ -50,11 +56,15 @@ export default function PdfViewer({
   stitchSettings: StitchSettings;
   filter: string;
   magnifying: boolean;
+  setPdfLoadStatus: Dispatch<SetStateAction<LoadStatusEnum>>;
+  setLineThicknessStatus: Dispatch<SetStateAction<LoadStatusEnum>>;
+  gridCenter: Point;
 }) {
   const [pageSizes, setPageSize] = useReducer(
     pageSizeReducer,
     new Map<number, { width: number; height: number }>(),
   );
+  const transformer = useTransformerContext();
 
   function onDocumentLoadSuccess(docProxy: PDFDocumentProxy) {
     const numPages = docProxy.numPages;
@@ -71,12 +81,22 @@ export default function PdfViewer({
 
   function onPageLoadSuccess(pdfProxy: PDFPageProxy) {
     const scale = (pdfProxy.userUnit || 1) * PDF_TO_CSS_UNITS;
+    const width = pdfProxy.view[2] * scale;
+    const height = pdfProxy.view[3] * scale;
     setPageSize({
       action: "setPage",
       pageNumber: pdfProxy.pageNumber,
-      width: pdfProxy.view[2] * scale,
-      height: pdfProxy.view[3] * scale,
+      width,
+      height,
     });
+    if (pageCount === 1) {
+      transformer.recenter(gridCenter, width, height);
+    }
+  }
+
+  function onPageRenderSuccess() {
+    setPdfLoadStatus(LoadStatusEnum.SUCCESS);
+    setLineThicknessStatus(LoadStatusEnum.SUCCESS);
   }
 
   const customTextRenderer = useCallback(({ str }: { str: string }) => {
@@ -138,7 +158,12 @@ export default function PdfViewer({
             >
               {value != 0 && (
                 <RenderContext.Provider
-                  value={{ erosions: lineThickness, layers, magnifying }}
+                  value={{
+                    erosions: lineThickness,
+                    layers,
+                    magnifying,
+                    onPageRenderSuccess,
+                  }}
                 >
                   <Page
                     scale={PDF_TO_CSS_UNITS}