+++ app/_hooks/use-render-context.ts
@@ -6,12 +6,14 @@ export interface RenderContextType {
   erosions: number;
   layers: Layers;
   magnifying: boolean;
+  onPageRenderSuccess: () => void;
 }
 
 export const RenderContext = createContext<RenderContextType>({
   erosions: 0,
   layers: {},
   magnifying: false,
+  onPageRenderSuccess: () => {},
 });
 
 export default function useRenderContext() {